requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Processing
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
pdfplumber==0.10.3
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.4
pyyaml==6.0.1
openpyxl==3.1.2

# Web Scraping
playwright==1.40.0
selenium==4.15.0
fake-useragent==1.4.0
requests==2.31.0
scrapy==2.11.0

# ML/AI
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
chromadb==0.4.22
qdrant-client==1.6.7
langchain==0.0.350
openai==1.3.5
spacy==3.7.2
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
elastic-apm==6.15.2

# Utilities
celery==5.3.4
firebase-admin==6.2.0
websockets==12.0
asyncio-mqtt==0.16.1